Update README.md
Lot of students aim to go for higher studies abroad. But the process is very confusing and expensive. 
Each decision about which college to apply  takes lot of consideration and the aim of our project
is to provide a comprehensive method to find suitable colleges for a person based on previous data.